+++ src/components/ContentPaper.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components/macro';
+import Paper from '@material-ui/core/Paper';
+
+const ContentPaper = styled(Paper)`
+  padding: 20px;
+`;
+
+export default ContentPaper;
