+++ src/pages/index.tsx
@@ -7,8 +7,7 @@ import Footer from 'components/Footer';
 const Books = lazy(() => import(/* webpackChunkName: 'books' */ './Books'));
 const Login = lazy(() => import(/* webpackChunkName: 'login' */ './Login'));
 const Top = lazy(() => import(/* webpackChunkName: 'top' */ './Top'));
-const TermsOfService = lazy(() => import(/* webpackChunkName: 'terms-of-service' */ './TermsOfService'));
-const PrivacyPolicy = lazy(() => import(/* webpackChunkName: 'privacy-policy' */ './PrivacyPolicy'));
+const Terms = lazy(() => import(/* webpackChunkName: 'terms' */ './Terms'));
 
 export default function RootRouter() {
   return (
@@ -17,8 +16,7 @@ export default function RootRouter() {
       <Switch>
         <Route path="/books" component={Books} />
         <Route path="/login" exact component={Login} />
-        <Route path="/terms-of-service" exact component={TermsOfService} />
-        <Route path="/privacy-policy" exact component={PrivacyPolicy} />
+        <Route path="/terms" component={Terms} />
         <Route component={Top} />
       </Switch>
       <Footer />
